feat: ログイン状態を保持するように実装
