+++ lib/state/user_state.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:diary_demo_app/domain/user_property.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
@@ -35,10 +37,10 @@ class UserState extends ChangeNotifier {
 
   /// ログイン処理
   Future<void> signIn(String mail, String password) async {
+    print('mail=$mail, password=$password');
     try {
       final credential = await FirebaseAuth.instance
           .signInWithEmailAndPassword(email: mail, password: password);
-      print(credential);
       final userId = credential.user?.uid ?? '';
       const userName = 'テストユーザー1';
       _userProperty = UserProperty.createAuthenticatedUser(
@@ -50,6 +52,8 @@ class UserState extends ChangeNotifier {
         print('No user found for that email.');
       } else if (e.code == 'wrong-password') {
         print('Wrong password provided for that user.');
+      } else {
+        print('サインイン時に原因不明のエラーが発生しました $e');
       }
     }
     notifyListeners();
@@ -60,5 +64,21 @@ class UserState extends ChangeNotifier {
     await FirebaseAuth.instance.signOut();
     _userProperty = UserProperty.createAnonymousUser();
     notifyListeners();
+  }
+
+  /// ログイン状態をチェックし、更新する
+  /// TODO: 1回実行するように改善する
+  Future<void> checkSignInStatus() async {
+    FirebaseAuth.instance.authStateChanges().listen((User? user) {
+      if (user == null) {
+        _userProperty = UserProperty.createAnonymousUser();
+        notifyListeners();
+      } else {
+        var userId = user.uid;
+        _userProperty = UserProperty.createAuthenticatedUser(
+            userId: userId, userName: 'テストユーザーhogehoge');
+        notifyListeners();
+      }
+    });
   }
 }
